+++ e2e-tests/development-runtime/gatsby-config.js
@@ -44,7 +44,7 @@ module.exports = {
       },
     },
     // this (optional) plugin enables Progressive Web App + Offline functionality
-    // To learn more, visit: https://gatsby.app/offline
+    // To learn more, visit: https://gatsby.dev/offline
     // 'gatsby-plugin-offline',
   ],
 }