+++ packages/gatsby-plugin-google-analytics/index.d.ts
@@ -4,8 +4,8 @@ interface OutboundLinkProps {
   onClick?: (event: React.MouseEvent<HTMLAnchorElement>) => void
 }
 
-class OutboundLink extends React.Component<
+export class OutboundLink extends React.Component<
   OutboundLinkProps & React.HTMLProps<HTMLAnchorElement>,
   any
 > {}
-export { OutboundLink }
+