+++ packages/gatsby-plugin-manifest/src/__tests__/gatsby-node.js
@@ -22,6 +22,7 @@ jest.mock(`sharp`, () => {
       }()
   )
   sharp.simd = jest.fn()
+  sharp.concurrency = jest.fn()
   return sharp
 })
 const fs = require(`fs`)