+++ www/src/components/ecosystem/ecosystem-featured-item.js
@@ -1,6 +1,6 @@
 import React from "react"
 import PropTypes from "prop-types"
-import styled from "react-emotion"
+import styled from "@emotion/styled"
 import { Link } from "gatsby"
 import Img from "gatsby-image"
 
@@ -17,7 +17,7 @@ const MAX_DESCRIPTION_LENGTH = 100
 const EcosystemFeaturedItemRoot = styled(HorizontalScrollerItem)`
   margin-right: ${rhythm(options.blockMarginBottom)};
 
-  ${presets.Tablet} {
+  ${presets.Md} {
     border-bottom: 1px solid ${colors.gray.superLight};
     box-shadow: none;
     margin: 0;
@@ -35,14 +35,14 @@ export const BlockLink = styled(Link)`
   height: 100%;
   padding: ${rhythm(3 / 4)};
 
-  ${presets.Tablet} {
+  ${presets.Md} {
     border-radius: 0;
     box-shadow: none;
     transition: all ${presets.animation.speedDefault}
       ${presets.animation.curveDefault};
   }
 
-  ${presets.Desktop} {
+  ${presets.Lg} {
     :hover {
       background: ${colors.ui.whisper};
     }